clean: restructured camera to be main draw-er in prep for serializing game objects
