+++ go.mod
@@ -5,5 +5,6 @@ go 1.15
 require (
 	github.com/ethereum/go-ethereum v1.9.24
 	github.com/golang/mock v1.4.4
+	github.com/hajimehoshi/ebiten v1.12.5
 	github.com/hajimehoshi/ebiten/v2 v2.1.0-alpha
 )
