+++ server/main.go
@@ -20,6 +20,12 @@ func main() {
 			fmt.Printf("couldn't accept connection from %v\n", conn.LocalAddr().String())
 		} else {
 			fmt.Printf("connection accepted from %v\n", conn.LocalAddr().String())
+
+			for i := 0; i < 10; i++ {
+				fmt.Fprintf(conn, "here is some stuff, please take %v\n", i+1)
+			}
+			fmt.Fprint(conn, "DONE\n")
+
 			conn.Close()
 		}
 	}
